+++ packages/core/src/connectivity/scp-cf/authorization-header.spec.ts
@@ -1,3 +1,4 @@
+import { createLogger } from '@sap-cloud-sdk/util';
 import {
   ODataGetAllRequestConfig,
   ODataRequest
@@ -12,13 +13,34 @@ import { Destination } from './destination';
 import {
   addAuthorizationHeader,
   buildAndAddAuthorizationHeader,
-  buildAuthorizationHeaders
+  getAuthHeaders
 } from './authorization-header';
 
-describe('buildAuthorizationHeaders', () => {
+const principalPropagationDestination = {
+  url: '',
+  authentication: 'PrincipalPropagation',
+  proxyType: 'OnPremise',
+  proxyConfiguration: {
+    headers: {
+      'SAP-Connectivity-Authentication': 'someValue',
+      'Proxy-Authorization': 'someProxyValue'
+    }
+  }
+} as Destination;
+
+function removeSapConnectivityAuthentication(destination) {
+  const destinationWithoutAuth = JSON.parse(JSON.stringify(destination));
+  delete destinationWithoutAuth!.proxyConfiguration!.headers![
+    'SAP-Connectivity-Authentication'
+  ];
+
+  return destinationWithoutAuth;
+}
+
+describe('getAuthHeaders', () => {
   describe('basic authentication', () => {
     it('creates basic authentication headers from destination credentials.', async () => {
-      const headers = await buildAuthorizationHeaders(defaultDestination);
+      const headers = await getAuthHeaders(defaultDestination);
       expect(headers.authorization).toBe(defaultBasicCredentials);
     });
   });
@@ -26,7 +48,7 @@ describe('buildAuthorizationHeaders', () => {
   describe('no authentication', () => {
     it('does not throw on NoAuthentication', async () => {
       await expect(
-        buildAuthorizationHeaders({
+        getAuthHeaders({
           url: 'https://example.com',
           authentication: 'NoAuthentication'
         })
@@ -35,12 +57,12 @@ describe('buildAuthorizationHeaders', () => {
 
     it('defaults to NoAuthentication', async () => {
       await expect(
-        buildAuthorizationHeaders({ url: 'https://example.com' })
+        getAuthHeaders({ url: 'https://example.com' })
       ).resolves.not.toThrow();
     });
 
     it('creates no authentication headers when only url is defined in a destination.', async () => {
-      const headers = await buildAuthorizationHeaders({
+      const headers = await getAuthHeaders({
         url: defaultDestination.url
       });
       expect(headers.authorization).toBeUndefined();
@@ -50,7 +72,7 @@ describe('buildAuthorizationHeaders', () => {
   describe('client credentials', () => {
     it('does not throw on ClientCertificateAuthentication', async () => {
       await expect(
-        buildAuthorizationHeaders({
+        getAuthHeaders({
           url: 'https://example.com',
           authentication: 'ClientCertificateAuthentication'
         })
@@ -59,28 +81,36 @@ describe('buildAuthorizationHeaders', () => {
   });
 
   describe('principal propagation', () => {
-    it('does not throw on Principal Propagation', async () => {
-      const destination = {
-        url: '',
-        authentication: 'PrincipalPropagation',
-        proxyType: 'OnPremise',
-        proxyConfiguration: {
-          headers: {
-            'SAP-Connectivity-Authentication': 'someValue',
-            'Proxy-Authorization': 'someProxyValue'
-          }
-        }
-      } as Destination;
-
-      const headers = await buildAuthorizationHeaders(destination);
+    it('builds headers', async () => {
+      const headers = await getAuthHeaders(principalPropagationDestination);
       checkHeaders(headers);
+    });
 
-      delete destination!.proxyConfiguration!.headers![
-        'SAP-Connectivity-Authentication'
-      ];
+    it('throws when no `SAP-Connectivity-Authentication` header is given', async () => {
       await expect(
-        buildAuthorizationHeaders(destination)
-      ).rejects.toThrowErrorMatchingSnapshot();
+        getAuthHeaders(
+          removeSapConnectivityAuthentication(principalPropagationDestination)
+        )
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        '"Principal propagation was selected in destination, but no SAP-Connectivity-Authentication bearer header was added by connectivity service."'
+      );
+    });
+
+    it('logs a warning instead of throwing an error for custom `SAP-Connectivity-Authentication` header', async () => {
+      const logger = createLogger('authorization-header');
+      const warnSpy = jest.spyOn(logger, 'warn');
+
+      const authHeader = { 'SAP-Connectivity-Authentication': 'token' };
+      await expect(
+        getAuthHeaders(
+          removeSapConnectivityAuthentication(principalPropagationDestination),
+          authHeader
+        )
+      ).resolves.toEqual(authHeader);
+
+      expect(warnSpy).toHaveBeenCalledWith(
+        'Found custom authorization headers. The given destination also provides authorization headers. This might be unintended. The custom headers from the request config will be used.'
+      );
     });
   });
 
@@ -103,7 +133,7 @@ describe('buildAuthorizationHeaders', () => {
         ]
       };
 
-      const actual = await buildAuthorizationHeaders(destination);
+      const actual = await getAuthHeaders(destination);
 
       expect(actual).toEqual({
         authorization: destination.authTokens![0].http_header.value
@@ -130,7 +160,7 @@ describe('buildAuthorizationHeaders', () => {
         ]
       };
 
-      const actual = await buildAuthorizationHeaders(destination);
+      const actual = await getAuthHeaders(destination);
 
       expect(actual).toEqual({
         authorization: 'Bearer some.token'
@@ -157,7 +187,7 @@ describe('buildAuthorizationHeaders', () => {
         ]
       };
 
-      const actual = await buildAuthorizationHeaders(destination);
+      const actual = await getAuthHeaders(destination);
 
       expect(actual).toEqual({
         authorization: 'Bearer some.token'
@@ -194,9 +224,13 @@ describe('buildAuthorizationHeaders', () => {
         ]
       };
 
-      await expect(
-        buildAuthorizationHeaders(destination)
-      ).rejects.toThrowErrorMatchingSnapshot();
+      await expect(getAuthHeaders(destination)).rejects
+        .toThrowErrorMatchingInlineSnapshot(`
+              "The destination tried to provide authorization tokens but failed in all cases. This is most likely due to misconfiguration.
+              Original error messages:
+              error
+              error"
+            `);
     });
 
     it('uses the first authToken where the error property is falsy', async () => {
@@ -226,7 +260,7 @@ describe('buildAuthorizationHeaders', () => {
           }
         ]
       };
-      const headers = await buildAuthorizationHeaders(destination);
+      const headers = await getAuthHeaders(destination);
       expect(headers.authorization).toBe('Bearer some.other.token');
     });
 
@@ -238,8 +272,10 @@ describe('buildAuthorizationHeaders', () => {
       };
 
       await expect(
-        buildAuthorizationHeaders(destination)
-      ).rejects.toThrowErrorMatchingSnapshot();
+        getAuthHeaders(destination)
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        '"AuthenticationType is \\"BasicAuthentication\\", but \\"username\\" and / or \\"password\\" are missing!"'
+      );
     });
 
     it('Throws an error if no authTokens are present, password is null and authenticationType is wrongly set', async () => {
@@ -250,8 +286,10 @@ describe('buildAuthorizationHeaders', () => {
       };
 
       await expect(
-        buildAuthorizationHeaders(destination)
-      ).rejects.toThrowErrorMatchingSnapshot();
+        getAuthHeaders(destination)
+      ).rejects.toThrowErrorMatchingInlineSnapshot(
+        '"AuthenticationType is \\"BasicAuthentication\\", but \\"username\\" and / or \\"password\\" are missing!"'
+      );
     });
   });
 });
@@ -329,7 +367,9 @@ describe('[deprecated]', () => {
     ];
     await expect(
       buildAndAddAuthorizationHeader(destination)({})
-    ).rejects.toThrowErrorMatchingSnapshot();
+    ).rejects.toThrowErrorMatchingInlineSnapshot(
+      '"Principal propagation was selected in destination, but no SAP-Connectivity-Authentication bearer header was added by connectivity service."'
+    );
   });
 });